fix(trusty-review): raise inference health-probe timeout for Bedrock cold-start (closes #739)
- Add `health_probe_timeout()` helper reading `TRUSTY_REVIEW_HEALTH_TIMEOUT_SECS`
  (default 10 s, was hard-coded 3 s); filters zero/invalid values to prevent
  accidental permanent-unknown state.
- Map probe timeout → `InferenceStatus::Unknown` (was `Unreachable`) so a slow
  Bedrock cold-start (~7.4 s observed) does not falsely report `degraded` health.
  Real review calls have a ~300 s budget; the probe budget is much shorter.
- Update `compute_status` to treat `Unknown` inference as non-degrading (same as
  `Ok`): only `Unreachable` / `AuthError` degrade top-level status.
- Split `inference_probe.rs` tests into `inference_probe_tests.rs` to satisfy the
  500-line cap (#610) after adding the new helper + four env-override tests.
- Update `InferenceProbe::default()` doc comment and `handle_health` doc to
  reflect the new configurable timeout and Unknown→non-degraded semantics.
