+++ crates/trusty-review/src/service/inference_probe_tests.rs
@@ -1,0 +1,345 @@
+//! Unit tests for `service::inference_probe`.
+//!
+//! Why: split from `inference_probe.rs` to keep that file under the 500-line
+//! cap (#610) while preserving full coverage of the probe logic, error mapping,
+//! cache TTL, and the configurable timeout helper (#739).
+//! What: exercises `InferenceStatus`, `map_llm_error`, `run_probe`,
+//! `InferenceProbe`, and `health_probe_timeout` via stub `LlmProvider` impls.
+//! Test: this is the test module; every function is a self-contained test.
+
+use super::*;
+use async_trait::async_trait;
+use std::sync::atomic::{AtomicU32, Ordering};
+
+// ── Stub providers ────────────────────────────────────────────────────────
+
+struct OkLlm;
+
+#[async_trait]
+impl LlmProvider for OkLlm {
+    fn name(&self) -> &str {
+        "ok-stub"
+    }
+
+    async fn complete(&self, req: LlmRequest) -> Result<LlmResponse, LlmError> {
+        Ok(LlmResponse {
+            text: "hi".to_string(),
+            model: req.model.clone(),
+            input_tokens: 1,
+            output_tokens: 1,
+            latency_ms: 0,
+            cost_usd: 0.0,
+        })
+    }
+}
+
+struct AuthErrorLlm;
+
+#[async_trait]
+impl LlmProvider for AuthErrorLlm {
+    fn name(&self) -> &str {
+        "auth-error-stub"
+    }
+
+    async fn complete(&self, _req: LlmRequest) -> Result<LlmResponse, LlmError> {
+        Err(LlmError::AccessDenied("invalid api key".into()))
+    }
+}
+
+struct TransportErrorLlm;
+
+#[async_trait]
+impl LlmProvider for TransportErrorLlm {
+    fn name(&self) -> &str {
+        "transport-stub"
+    }
+
+    async fn complete(&self, _req: LlmRequest) -> Result<LlmResponse, LlmError> {
+        Err(LlmError::Transport("connection refused".into()))
+    }
+}
+
+struct HungLlm;
+
+#[async_trait]
+impl LlmProvider for HungLlm {
+    fn name(&self) -> &str {
+        "hung-stub"
+    }
+
+    async fn complete(&self, _req: LlmRequest) -> Result<LlmResponse, LlmError> {
+        // Simulate an endpoint that never responds within the probe timeout.
+        tokio::time::sleep(Duration::from_secs(60)).await;
+        Err(LlmError::Transport("hung".into()))
+    }
+}
+
+/// A counting stub to verify the cache prevents redundant probes.
+struct CountingLlm {
+    calls: Arc<AtomicU32>,
+}
+
+#[async_trait]
+impl LlmProvider for CountingLlm {
+    fn name(&self) -> &str {
+        "counting-stub"
+    }
+
+    async fn complete(&self, req: LlmRequest) -> Result<LlmResponse, LlmError> {
+        self.calls.fetch_add(1, Ordering::Relaxed);
+        Ok(LlmResponse {
+            text: "x".into(),
+            model: req.model.clone(),
+            input_tokens: 1,
+            output_tokens: 1,
+            latency_ms: 0,
+            cost_usd: 0.0,
+        })
+    }
+}
+
+// ── InferenceStatus tests ─────────────────────────────────────────────────
+
+#[test]
+fn probe_status_serialises_lowercase() {
+    assert_eq!(
+        serde_json::to_string(&InferenceStatus::Ok).unwrap(),
+        "\"ok\""
+    );
+    assert_eq!(
+        serde_json::to_string(&InferenceStatus::Unreachable).unwrap(),
+        "\"unreachable\""
+    );
+    assert_eq!(
+        serde_json::to_string(&InferenceStatus::AuthError).unwrap(),
+        "\"auth_error\""
+    );
+    assert_eq!(
+        serde_json::to_string(&InferenceStatus::Unknown).unwrap(),
+        "\"unknown\""
+    );
+}
+
+#[test]
+fn probe_status_is_ok() {
+    assert!(InferenceStatus::Ok.is_ok());
+    assert!(!InferenceStatus::Unreachable.is_ok());
+    assert!(!InferenceStatus::AuthError.is_ok());
+    assert!(!InferenceStatus::Unknown.is_ok());
+}
+
+// ── Error mapping tests ───────────────────────────────────────────────────
+
+#[test]
+fn error_mapping_access_denied_is_auth_error() {
+    let status = map_llm_error(&LlmError::AccessDenied("denied".into()));
+    assert_eq!(status, InferenceStatus::AuthError);
+}
+
+#[test]
+fn error_mapping_model_not_found_is_auth_error() {
+    let status = map_llm_error(&LlmError::ModelNotFound("no-model".into()));
+    assert_eq!(status, InferenceStatus::AuthError);
+}
+
+#[test]
+fn error_mapping_model_not_ready_is_auth_error() {
+    let status = map_llm_error(&LlmError::ModelNotReady("creating".into()));
+    assert_eq!(status, InferenceStatus::AuthError);
+}
+
+#[test]
+fn error_mapping_validation_is_auth_error() {
+    let status = map_llm_error(&LlmError::Validation("bad prefix".into()));
+    assert_eq!(status, InferenceStatus::AuthError);
+}
+
+#[test]
+fn error_mapping_transport_is_unreachable() {
+    let status = map_llm_error(&LlmError::Transport("connection refused".into()));
+    assert_eq!(status, InferenceStatus::Unreachable);
+}
+
+#[test]
+fn error_mapping_rate_limited_is_unreachable() {
+    let status = map_llm_error(&LlmError::RateLimited);
+    assert_eq!(status, InferenceStatus::Unreachable);
+}
+
+#[test]
+fn error_mapping_upstream_5xx_is_unreachable() {
+    let status = map_llm_error(&LlmError::Upstream {
+        status: 503,
+        body: "overloaded".into(),
+    });
+    assert_eq!(status, InferenceStatus::Unreachable);
+}
+
+// ── Live probe tests ──────────────────────────────────────────────────────
+
+#[tokio::test]
+async fn probe_returns_ok_on_success() {
+    let llm: Arc<dyn LlmProvider> = Arc::new(OkLlm);
+    let status = run_probe(&llm, "test-model", Duration::from_secs(5)).await;
+    assert_eq!(status, InferenceStatus::Ok);
+}
+
+#[tokio::test]
+async fn probe_returns_auth_error_on_access_denied() {
+    let llm: Arc<dyn LlmProvider> = Arc::new(AuthErrorLlm);
+    let status = run_probe(&llm, "test-model", Duration::from_secs(5)).await;
+    assert_eq!(status, InferenceStatus::AuthError);
+}
+
+#[tokio::test]
+async fn probe_returns_unreachable_on_transport() {
+    let llm: Arc<dyn LlmProvider> = Arc::new(TransportErrorLlm);
+    let status = run_probe(&llm, "test-model", Duration::from_secs(5)).await;
+    assert_eq!(status, InferenceStatus::Unreachable);
+}
+
+#[tokio::test(start_paused = true)]
+async fn probe_timeout_returns_unknown() {
+    // The HungLlm sleeps 60 s; the probe timeout is 10 ms.
+    // With paused clock, `tokio::time::sleep` returns instantly when
+    // the runtime advances time; `timeout` fires before HungLlm wakes.
+    // Timeout → `Unknown` (not `Unreachable`) so a slow Bedrock cold-start
+    // does not falsely degrade health (#739).
+    let llm: Arc<dyn LlmProvider> = Arc::new(HungLlm);
+    let status = run_probe(&llm, "test-model", Duration::from_millis(10)).await;
+    assert_eq!(
+        status,
+        InferenceStatus::Unknown,
+        "probe timeout must return Unknown, not Unreachable (#739)"
+    );
+}
+
+// ── Cache TTL tests ───────────────────────────────────────────────────────
+
+#[tokio::test]
+async fn probe_cache_prevents_redundant_calls() {
+    let calls = Arc::new(AtomicU32::new(0));
+    let llm: Arc<dyn LlmProvider> = Arc::new(CountingLlm {
+        calls: Arc::clone(&calls),
+    });
+
+    // Long TTL: 60 s — the cache should remain warm across both calls.
+    let probe = InferenceProbe::new(Duration::from_secs(60), Duration::from_secs(5));
+
+    let s1 = probe.probe(&llm, "m").await;
+    let s2 = probe.probe(&llm, "m").await;
+
+    assert_eq!(s1, InferenceStatus::Ok);
+    assert_eq!(s2, InferenceStatus::Ok);
+    assert_eq!(
+        calls.load(Ordering::Relaxed),
+        1,
+        "provider must be called exactly once when cache is warm"
+    );
+}
+
+#[tokio::test(start_paused = true)]
+async fn probe_cache_ttl_zero_always_reprobes() {
+    let calls = Arc::new(AtomicU32::new(0));
+    let llm: Arc<dyn LlmProvider> = Arc::new(CountingLlm {
+        calls: Arc::clone(&calls),
+    });
+
+    // TTL of 0 means the cache always looks stale.
+    let probe = InferenceProbe::new(Duration::ZERO, Duration::from_secs(5));
+
+    probe.probe(&llm, "m").await;
+    probe.probe(&llm, "m").await;
+
+    assert_eq!(
+        calls.load(Ordering::Relaxed),
+        2,
+        "zero TTL must reprobe on every call"
+    );
+}
+
+// ── health_probe_timeout() tests ──────────────────────────────────────────
+
+/// Returns 10 s when the env var is absent.
+///
+/// Why: verifies the documented default so operators know what to expect
+/// without setting the variable.
+/// What: calls `health_probe_timeout()` with the env var unset.
+/// Test: this test (serial to prevent env-var races with sibling tests).
+#[test]
+#[serial_test::serial]
+fn health_probe_timeout_default() {
+    // SAFETY: serial_test::serial ensures no other thread mutates env vars
+    // concurrently in this process during this test.
+    unsafe { std::env::remove_var("TRUSTY_REVIEW_HEALTH_TIMEOUT_SECS") };
+    let t = health_probe_timeout();
+    assert_eq!(
+        t,
+        Duration::from_secs(10),
+        "default probe timeout must be 10 s (#739)"
+    );
+}
+
+/// Returns the caller-supplied value when the env var is a valid non-zero u64.
+///
+/// Why: verifies that operators can raise or lower the timeout without
+/// recompiling.
+/// What: sets `TRUSTY_REVIEW_HEALTH_TIMEOUT_SECS=15` and checks the result.
+/// Test: this test.
+#[test]
+#[serial_test::serial]
+fn health_probe_timeout_env_override() {
+    // SAFETY: serial_test::serial ensures no other thread mutates env vars
+    // concurrently in this process during this test.
+    unsafe { std::env::set_var("TRUSTY_REVIEW_HEALTH_TIMEOUT_SECS", "15") };
+    let t = health_probe_timeout();
+    unsafe { std::env::remove_var("TRUSTY_REVIEW_HEALTH_TIMEOUT_SECS") };
+    assert_eq!(
+        t,
+        Duration::from_secs(15),
+        "env-var override must be honoured"
+    );
+}
+
+/// Falls back to 10 s when the env var contains a non-numeric value.
+///
+/// Why: a mis-typed env var (e.g. `TRUSTY_REVIEW_HEALTH_TIMEOUT_SECS=ten`)
+/// must not panic or produce an unintended value; fallback to default is safest.
+/// What: sets an invalid value and asserts the default is used.
+/// Test: this test.
+#[test]
+#[serial_test::serial]
+fn health_probe_timeout_env_invalid_falls_back() {
+    // SAFETY: serial_test::serial ensures no other thread mutates env vars
+    // concurrently in this process during this test.
+    unsafe { std::env::set_var("TRUSTY_REVIEW_HEALTH_TIMEOUT_SECS", "not-a-number") };
+    let t = health_probe_timeout();
+    unsafe { std::env::remove_var("TRUSTY_REVIEW_HEALTH_TIMEOUT_SECS") };
+    assert_eq!(
+        t,
+        Duration::from_secs(10),
+        "invalid env var must fall back to 10 s default"
+    );
+}
+
+/// Falls back to 10 s when the env var is zero (prevents a zero timeout).
+///
+/// Why: a zero timeout would make every probe immediately time out and return
+/// `Unknown`, rendering the inference field permanently `unknown`.  Treating
+/// zero as "use default" is a safe guard against misconfiguration.
+/// What: sets `TRUSTY_REVIEW_HEALTH_TIMEOUT_SECS=0` and asserts the default.
+/// Test: this test.
+#[test]
+#[serial_test::serial]
+fn health_probe_timeout_env_zero_falls_back() {
+    // SAFETY: serial_test::serial ensures no other thread mutates env vars
+    // concurrently in this process during this test.
+    unsafe { std::env::set_var("TRUSTY_REVIEW_HEALTH_TIMEOUT_SECS", "0") };
+    let t = health_probe_timeout();
+    unsafe { std::env::remove_var("TRUSTY_REVIEW_HEALTH_TIMEOUT_SECS") };
+    assert_eq!(
+        t,
+        Duration::from_secs(10),
+        "zero env var must fall back to 10 s default"
+    );
+}
